Solved Nortia Racing problem
Instead of using backtracking, it's better to just use the math formula ax + by. Backtracking works. Just it is VERY slow. We can bring our time complexity from exponential to O(n) with this formula. Backtracking isn't the solution because the order the buttons are pressed does not matter. We just want to know the distances we could travel regardless of the button pushes.